+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix fetching `AWSClusterRoleIdentity`.
+- Remove `k8s` from the domain name.
+
 ## [0.2.2] - 2022-08-04
 
 ## [0.2.1] - 2022-03-24